Refactor to event repository
