+++ lib/business_logic/item_action.dart
@@ -8,13 +8,12 @@ import 'package:spare_parts/business_logic/borrowing_request_dialog.dart';
 import 'package:spare_parts/entities/custom_user.dart';
 import 'package:spare_parts/entities/event.dart';
 import 'package:spare_parts/entities/inventory_item.dart';
-import 'package:spare_parts/services/firestore_service.dart';
 import 'package:spare_parts/utilities/constants.dart';
 import 'package:spare_parts/utilities/helpers.dart';
-import 'package:spare_parts/widgets/inputs/user_selection_dialog.dart';
-import 'package:spare_parts/widgets/inventory_list_item/inventory_item_form.dart';
 import 'package:spare_parts/widgets/dialogs/print_dialog/print_dialog_mobile.dart'
     if (dart.library.html) 'package:spare_parts/widgets/dialogs/print_dialog/print_dialog_web.dart';
+import 'package:spare_parts/widgets/inputs/user_selection_dialog.dart';
+import 'package:spare_parts/widgets/inventory_list_item/inventory_item_form.dart';
 
 import '../services/repositories/repositories.dart';
 
@@ -125,9 +124,9 @@ class BorrowItemAction extends ItemAction {
 
   @override
   handle(BuildContext context, InventoryItem item) {
-    final firestoreService = context.read<FirestoreService>();
     final inventoryItemRepository = context.read<InventoryItemRepository>();
     final borrowingRuleRepository = context.read<BorrowingRuleRepository>();
+    final eventRepository = context.read<EventRepository>();
     final borrowingRequestRepository =
         context.read<BorrowingRequestRepository>();
     final auth = context.read<FirebaseAuth>();
@@ -179,7 +178,7 @@ class BorrowItemAction extends ItemAction {
           type: 'Borrow',
           createdAt: DateTime.now(),
         );
-        await firestoreService.addEvent(item.id, event);
+        await eventRepository.addEvent(item.id, event);
 
         return true;
       },
@@ -201,7 +200,7 @@ class ReleaseItemAction extends ItemAction {
 
   @override
   handle(BuildContext context, InventoryItem item) {
-    final firestoreService = context.read<FirestoreService>();
+    final eventRepository = context.read<EventRepository>();
     final inventoryItemRepository = context.read<InventoryItemRepository>();
     final auth = context.read<FirebaseAuth>();
 
@@ -213,7 +212,7 @@ class ReleaseItemAction extends ItemAction {
           type: 'Release',
           createdAt: DateTime.now(),
         );
-        await firestoreService.addEvent(item.id, event);
+        await eventRepository.addEvent(item.id, event);
         await inventoryItemRepository.release(item);
 
         return true;
