+++ lib/services/firestore_service.dart
@@ -20,22 +20,4 @@ class FirestoreService {
   CollectionReference get metaCollection => _firestore.collection('meta');
 
   CollectionReference get usersCollection => _firestore.collection('users');
-
-
-  Stream<List<Event>> getEventsStream(String? inventoryItemId) {
-    return itemsCollection
-        .doc(inventoryItemId)
-        .collection('events')
-        .orderBy('createdAt', descending: true)
-        .snapshots()
-        .map((snap) =>
-            snap.docs.map((doc) => Event.fromFirestore(doc)).toList());
-  }
-
-  Future<void> addEvent(String inventoryItemId, Event event) async {
-    await itemsCollection
-        .doc(inventoryItemId)
-        .collection('events')
-        .add(event.toFirestore());
-  }
 }
