+++ lib/services/repositories/event_repository.dart
@@ -1,0 +1,23 @@
+import 'package:spare_parts/entities/event.dart';
+import 'package:spare_parts/services/firestore_service.dart';
+
+class EventRepository extends FirestoreService {
+  EventRepository(super.firestore);
+
+  Stream<List<Event>> getEventsStream(String? inventoryItemId) {
+    return itemsCollection
+        .doc(inventoryItemId)
+        .collection('events')
+        .orderBy('createdAt', descending: true)
+        .snapshots()
+        .map((snap) =>
+            snap.docs.map((doc) => Event.fromFirestore(doc)).toList());
+  }
+
+  Future<void> addEvent(String inventoryItemId, Event event) async {
+    await itemsCollection
+        .doc(inventoryItemId)
+        .collection('events')
+        .add(event.toFirestore());
+  }
+}
