+++ lib/services/repositories/repositories.dart
@@ -3,3 +3,4 @@ export 'borrowing_rule_repository.dart';
 export 'borrowing_request_repository.dart';
 export 'meta_repository.dart';
 export 'user_repository.dart';
+export 'event_repository.dart';
